Add PDF viewing functionality
A new activity was added to view the PDF summaries. If this commit is
big, it's probably because I also refactored some of the logic behind
requesting the web pages.
There might be a problem building because I added a few Apache libraries
that don't seem to register correctly at runtime. Will fix soon.