Improve UI and app logic significantly
This commit makes the schedule update daily, adds support for reading the news off of the user's school's website, adds new art for the background of the class note buttons, formats the hard-coded UI into an adaptable one, improves login security and fixes a bunch of tiny warnings that were popping up everywhere. Also, the web client dependency was changed in favor of something more modern that was compatible with the latest Android SDK version. Overall, a seriously huge (and rewarding) commit!